Making sure generated C names have underscores representing namespace
When defining variables and functions inside a namespace, those variables and functions as generated in C need to have the namespace name prepended to them, to avoid name clashes with other namespace code.  Came up with a good way to do this by adding the concept of a "cname" that is associated with all identifiers.  These cnames are generated the first time they are needed and then cached for subsequent access.

Also, I renamed the Symbol Table to the Identifier Table, because that's what it is, and the name "Symbol Table" has always confused me, as it sounds like it's storing much more than it really is.