Cleaning up C code formatting
There were some problems with the way we were formatting the generated C code.  I fixed most of these problems by changing the way we handle tab levels.  The previous incarnation worked by tabbing once for each level up the AST.  This is not optimal, though, because node depth in an AST is often far deeper than structural node depth of the resulting C-code.  For this reason, we now to tab level management a bit more manually... within each visitor function as appropriate.

There are still some bugs here and there (such as tabbing being off when something is a single statement rather than a statement list), but we'll fix more as we go along.