Replaced existing identifier table system with a nested scoping system
Lists of identifiers now live in "scopes", and each scope can have a parent scope.  Identifier lookup is done in the usual way, with lookup errors causing the next higher scope to be examined, until finally an identifier not found error is given.

I also fixed quite a few memory leaks in the compiler.